order functionality was created but not finished

+++ wlm-ecommerce-product-service/src/main/java/com/ecommerce/wlm_ecommerce_product_service/application/service/OrderService.java
@@ -1,24 +1,25 @@
 package com.ecommerce.wlm_ecommerce_product_service.application.service;
 
+import com.ecommerce.wlm_ecommerce_product_service.domain.exception.ProductNotFoundException;
+import com.ecommerce.wlm_ecommerce_product_service.domain.exception.UserNotFoundException;
 import com.ecommerce.wlm_ecommerce_product_service.domain.model.Order;
 import com.ecommerce.wlm_ecommerce_product_service.domain.model.OrderItem;
+import com.ecommerce.wlm_ecommerce_product_service.domain.model.StatusPayment;
 import com.ecommerce.wlm_ecommerce_product_service.domain.repository.OrderRepository;
-import com.ecommerce.wlm_ecommerce_product_service.infrastructure.persistence.entity.OrderEntity;
-import com.ecommerce.wlm_ecommerce_product_service.infrastructure.persistence.entity.OrderItemsEntity;
 import com.ecommerce.wlm_ecommerce_product_service.infrastructure.persistence.entity.ProductEntity;
 import com.ecommerce.wlm_ecommerce_product_service.infrastructure.persistence.repository.SpringDataProduct;
 import com.ecommerce.wlm_ecommerce_product_service.infrastructure.persistence.repository.SpringDataUser;
 import jakarta.transaction.Transactional;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
 import java.util.List;
-import java.util.NoSuchElementException;
+import java.util.stream.Collectors;
 
 @Service
 public class OrderService {
     private final OrderRepository orderRepository;
     private final SpringDataProduct springDataProduct;
+    private final SpringDataUser springDataUser;
 
     public OrderService(
             OrderRepository orderRepository,
@@ -26,42 +27,61 @@ public class OrderService {
             SpringDataUser springDataUser) {
         this.orderRepository = orderRepository;
         this.springDataProduct = springDataProduct;
+        this.springDataUser = springDataUser;
     }
 
     @Transactional
-    public OrderEntity createOrder(Order order){
-        OrderEntity orderEntity = new OrderEntity(
-                order.getStatusOrder(),
-                order.getStatusPayment(),
-                order.getUserId()
-        );
-
-        List<OrderItemsEntity> items = new ArrayList<>();
-        for(OrderItem item : order.getItems()){
-            ProductEntity product = springDataProduct.findById(item.getProductId())
-                    .orElseThrow(() -> new NoSuchElementException("Product not found"));
-
-            OrderItemsEntity itemsEntity = new OrderItemsEntity();
-            itemsEntity.setProductEntity(product);
-            itemsEntity.setQuantity(item.getQuantity());
-            itemsEntity.setOrder(orderEntity); // binding
-            items.add(itemsEntity);
+    public Order createOrder(Order orderRequest){
+        if(orderRequest.getUserId() <= 0 || springDataUser.findById(orderRequest.getUserId()).isEmpty()){
+            throw new UserNotFoundException(orderRequest.getUserId());
         }
 
-        orderEntity.setItems(items);
-        return orderRepository.save(orderEntity);
+        List<OrderItem> enrichedItems = orderRequest.getItems().stream()
+                .map(item -> {
+                    ProductEntity product = springDataProduct.findById(item.getProductId())
+                            .orElseThrow(() -> new ProductNotFoundException("Product not found: " + item.getProductId()));
+
+                    return new OrderItem(
+                            item.getQuantity(),
+                            null, // orderId not exists yet
+                            item.getProductId(),
+                            product.getPrice()
+                    );
+                }).collect(Collectors.toList());
+
+        return orderRepository.save(new Order(orderRequest.getUserId(), enrichedItems));
     }
 
-    public OrderEntity getOrderById(Long orderId){
-       return orderRepository.findOrderById(orderId);
+    public Order getOrderById(Long orderId){
+       if(orderId <= 0){
+           throw new IllegalArgumentException("orderId must be greater than zero");
+       }
+
+        return orderRepository.findOrderById(orderId);
     }
 
-    public List<OrderEntity> getAllOrders(Long userId){
-        return orderRepository.findOrdersByUserId(userId);
+    public List<Order> getAllOrders(Long userId){
+        if(userId <= 0){
+            throw new IllegalArgumentException("userId must be greater than zero");
+        }
+        return orderRepository.findAllByUserId(userId);
+    }
+
+    public Order changingPaymentStatus(Long orderId, StatusPayment statusPayment){
+        Order order = orderRepository.findOrderById(orderId);
+        order.changeStatusPayment(statusPayment);
+        return orderRepository.save(order);
     }
 
     @Transactional
-    public void deleteOrderById(Long orderId){
+    public void deleteOrder(Long userId, Long orderId){
+        if(orderId <= 0){
+            throw new IllegalArgumentException("Order id must to be greater than 0");
+        }
+
+        springDataUser.findById(userId)
+                .orElseThrow(UserNotFoundException::new);
+
         orderRepository.deleteOrderById(orderId);
     }
